Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

(1 inline comment)

This just provides FPU support if we need it.  Given that we've taken the time to figure it out, let's not lose the code.  We can #ifdef it out with CONFIG_FPU as you suggest.

The temperature sensor code doesn't currently implement the multi-point temporal correction algorithm that TI recommends for the TMP006.  That adds more complexity to the algorithm, making an already-somewhat-unwieldy fixed-point implementation even harder to understand/debug compared with a few calls to sqrtf() and the like.

Patch-set: 2
Label: Code-Review=+1
Label: Commit-Queue=0
Label: Verified=0