stm32l: add simple I2C slave driver
The I2C slave 8-bit address is 0xEC.
(so the real 7-bit I2C address is 0x76)
It implements reading and writing a 8-bit value at a 8-bit index.

It's just a dummy driver printing the writes and returning the inverted
index on reads.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=None
TEST=on ADV board, do I2C reads and writes from U-Boot running on the
AP, and see the writes displayed on the EC console and the reads
returning the right values.

Change-Id: Ic8e166535df632c9e4d939c687c5fba3aa3ec2d0