stm32l: allow to power on without XPSHOLD signal.
If we timed-out waiting for XPSHOLD, just power on the system (and disable
the power off triggered by XPSHOLD = 0).
This allows to run the USB loader and JTAG only software which don't set
the XPSHOLD signal, but it's a bit less safer since we might consume
power forever while the APU is in a non-working state.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=None
TEST=On ADV board, check we can do a full power on and an 8s power off
with and without XPSHOLD.

Change-Id: Ic83dde0bae61b6e4d9916f6bc09e6ceed5f491a5
Signed-off-by: Simon Glass <sjg@chromium.org>