Handle ghost key issue.
This is more complex than we thought before. We used to think of the case
that the existing 2 keys in key_down[] sit on diagonal and the incoming
key triggers a ghost key on the other side of the square. However, we
forgot to consider the situation that existing 2 keys sit on the
same row (or the same column), then the 3rd key sharing a same column
(or a row) triggers a ghost key.

More complex situation, there are 3 keys already existed in key_down[].
2 on the same row (or column) and the other one is alone. The incoming
4th key shares the same column (or row) with the alone key. In this case,
we could trigger 3 ghost key at maximum.

So, the better solution is to try every 3-key combination (the new key
and 2 keys existed in key_down[]). If any 3-key could generate a ghost key,
then ignore the new key.

BUG=none
TEST=on the bds.

Change-Id: I17e6e90d1f5e5fa71d31c31b231ca9ac449ed6b1