Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This at least cuts number of parameters from (# of sensors) * (# of thresholds) to (# of sensor types) * (# of thresholds).

Which is a help when we have 4 case sensors and 4 board sensors.

The I2C temp sensors next to the memory and PCH are still board sensors.  

We could also claim the EC is a board sensor, since it's not measuring CPU or case.  That'll further cut the memory usage.

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1
Label: Verified=0