Update patch set 1
Patch Set 1:

Hm... this is not actually for write protect. This is to give flashrom a hint which part of EC flash is the active image so that he can come back updating after jump. Even the RW part would be protected by this feature too.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0