Split lid handling into separate functions and add command interface
This has been useful for me to be able to test lid behavior remotely
since it is not available via servo.

This also has a minor change to send a task message after sending
the power button pulse so the state machine behaves properly.

BUG=none
TEST=Execute 'lidclose' and 'lidopen' commands via ec uart and
see the appropriate events set and wake behavior when the system
is off.  With a (not yet published) coreboot I am able to handle
lid close events to enter suspend.

Change-Id: Iec1c68121d42b66305ba5dfd20e81453538a97e2
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>