daisy v1.02: setup GPIO signal to audio codec
Set up signal to audio codec as open-drain output. This should
be benign on older revisions of Daisy.

The signal CODEC_INT is driven low whenever a keypress has been
processed. For Daisy, a delay of about ~1ms from the time a key
is pressed until the time the codec is signaled is okay. This
gives us time to scan the columns to see if a key was actually
pressed so that we don't cause spurious codec interrupts.

Signed-off-by: David Hendricks <dhendrix@chromium.org>

BUG=None
TEST=Tested on Daisy using oscilloscope

Change-Id: Id3564f4aacbf7294b7151b082075f3c3ec8b1eb2