add more explicit GPIO types
The current GPIO types cover many useful combinations, but there are
some used for stm32l boards which are not covered. Examples include
inputs where the internal pull-up resistor is activated and outputs
that are open-drain.

For now, the new types are simply kludged in to avoid effecting
functinality with the old types.

Signed-off-by: David Hendricks <dhendrix@chromium.org>

BUG=none
TEST=compile tested

Change-Id: I1a0c2e8366f6a82cd9cd7e83e57122944f2bdc2d