Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

It wasn't SPI that stopped working - it was flashing of U-Boot!

We do need to get to the bottom of this. I submitted something today which might help - we should retest and sort this out before submitting this.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=0