Update patch set 1
Patch Set 1:

Note: This came up when Luigi and I were updating the mkbp driver in the kernel. Basically, we figured it would be nice to be able to compare the total bytes sent across the wire with the number of 16-bit length field in the packet. It also seems more sensible for other calculations:

bytes in message payload = length - MSG_PROTO_BYTES and 

bytes involved in checksum = length - MSG_TRAILER_BYTES

Patch-set: 1
Label: Code-Review=0