WIP: add special version command to messaging protocol
This adds a special case for the host to query the messaging protocol
version. Since the protocol version must be known in order to decode
a message, the version itself is transferred in raw format with a
fixed-length and format.

Note: I'm open to suggestions on exactly how the version should look,
what each byte means exactly, if it should be a struct, etc.

BUG=none
TEST=compile tested only for now

Signed-off-by: David Hendricks <dhendrix@chromium.org>

Change-Id: I855173961ac4b9a090645d6b469ce6bf691dbedb