stm32: change slave address to 0x3c
This is a hack to avoid issues caused by incompatible
messaging protocol updates.

During protocol development, the length of a packet changed which
could cause the system to hang (or other issues) if the host
requested the wrong number of bytes from the EC. This avoids the
issue with development versions of the protocol, by simply making
the EC unresponsive on the old port.

BUG=none
TEST=Tested on Daisy 1.02 and EVT1

Change-Id: I96495d4c2bd14b377bef862801934d5168cb6cc7
Signed-off-by: David Hendricks <dhendrix@chromium.org>