Allow console commands abbreviation
The EC console input handling code is being enhanced to accept
abbreviated command names. If the abbreviation is unique, the
appropriate command is used, if the abbreviation is ambiguous, the
command is handled as nonexistent. The error message is being modified
to mention that the command either does not exist or is ambiguous.

BUG=none
TEST=manual
 . program the new EC image. Try entering at the console:
  > h
  Command 'h' either not found or ambiguous.
  Command returned error 1
  > he
  Known commands:
    adc            autofan        battery        ch             charger
  ...
  > help
  Known commands:
    adc            autofan        battery        ch             charger
  ...

Change-Id: Iaa3e91e1504e42daefb02d561e00c39003548197
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>