stm32: add a FIFO for keyboard usage
This CL adds a FIFO and a task for interrupting the AP to notify it
if FIFO entries exist. For now it is only used to store keystrokes
but could eventually be made more generic.

The change was prompted by the realization that the AP can be very
slow at scheduling the task which queries the EC for keystroke data.
The actual transfer of data is fast, but the work might get preempted
for hundreds millisecond.

This CL includes:
- Keyboard FIFO and modifications to Daisy/Snow to use it.
- FIFO task that wakes up periodically. If there is something in the
  FIFO then it will interrupt the host.
- Console commands for debugging

The added console commands are:
kb_status: print keyboard FIFO variables and content
kb_clear: clear keyboard FIFO entries
kb_inject <row> <col> <num_presses>: Simulate key presses

BUG=chrome-os-partner:9594
TEST=tested on daisy
Signed-off-by: David Hendricks <dhendrix@chromium.org>

Testing was mostly subjectively done by typing a sentence quickly
at the Chrome UI. kb_inject() was also used to rapidly fill the FIFO
to test overflow cases.

Change-Id: I6faff41b0f16b75d6426df13a239f84eb62634ec