Update patch set 7
Patch Set 7: I would prefer that you didn't submit this

I don't think this is the right solution to the problem.
I agree we need buffering, but
 why do we sometimes repeatedly resend an interrupt ?
 (the AP should not loose IT, no matter how it is)
 why do we need a separate task for the FIFO (and all the fancy locking) ?

The retrieval whole mechanism might not be efficient when we have several key strokes already queued.

As a side note, I'm updating the host messaging system to have the same code on Daisy and Link, WIP is there :
https://gerrit.chromium.org/gerrit/#change,22504

Patch-set: 7
Label: Code-Review=-1
Label: Commit-Queue=0
Label: Verified=0