Update patch set 8
Patch Set 8:

Regarding the protocol: we considered various other options, including sending keyboard events instead of keyboard states, compressing the states, and sending multiple messages at once.  In the end we opted for the simplest change that we could make, because of time constraints.  This change didn't require any kernel changes, so it was easier to test.  The intention is to improve this.

David: it is likely that the scheduling of additional work functions (one per keyboard state) is worsening the latency.  The kernel could easily keep a count of how many interrupts it has received, and request that many states in a single work function invocation.

Would you like to try that tomorrow?

Patch-set: 8
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0