Make host_command_received() store its parameters
When a command is queued up, we know where it is and may know how big its
parameters are. We certainly know about the amount of space available for
the response.

Rather than ask for this information again later, using host_get_buffer(),
just record it when the command is queued. This removes a somewhat
convoluted command processing mechanism, and permits us to check the
maximum buffer size available for the response if we want to.

BUG=chrome-os-partner:10535
TEST=manual:
build for all boards
build and boot on daisy
see that keyboard still works
build and boot on link
see that command processing still works

Change-Id: I465df66acb8df9e78fe7c2fd4b0d3152826a6704
Signed-off-by: Simon Glass <sjg@chromium.org>