Remove host_get_buffer()
We don't need this really, since the driver knows exactly where the
buffer is anyway, and the host_command layer is now told where it is
explicitly. Punt it.

BUG=chrome-os-partner:10535
TEST=manual:
build for all boards
build and boot on daisy
see that keyboard still works
build and boot on link
see that command processing still works

Change-Id: I21b72bc64c473f46a560eb2bd88e8ebac8f9f898
Signed-off-by: Simon Glass <sjg@chromium.org>