Update patch set 1
Patch Set 1: Verified; Ready

For the sake of expediency I'll go ahead and submit. We can re-visit if needed.

I think the problem is that the only facility we have for obtaining that info wasn't giving us what we really wanted. As you point out, the hardware can only write data to flash in half-words. However, the amount of data that should be transferred from the host per transaction is a different matter. I think it's useful to probe that info but we can just hardcode it on the flashrom side if there's a strong preference.

Patch-set: 1
Reviewer: Gerrit User 1000253 <1000253@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=+1
Label: Verified=+1