To optimize the flash read/write speed, add a BURST_INFO command.
This is different to the FLASH_INFO command, which provides the minimum
lengths of write/erase/protect command. This command is to provide
the maxmium lengths of blob in a read/write command.

I don't have daisy/lucas board on hand for testing, but the code looks
good to me.

BUG=None
TEST=Tested on link board. Speeds up 3 times.
Signed-off-by: Louis Yung-Chieh Lo <yjlou@chromium.org>

Change-Id: I16574c9c82b05979251af35058fbccfede2bce50