Update patch set 3
Patch Set 3:

sort-of.

Both lm4 and stm32 need some sort of watchdog_check() API.

The lm4 uses it to task_disable_irq()

The stm32 uses it to update its WWDG counter.

So we can't just call watchdog_trace() anymore.

Given this, it seems that the irq handler can be the same for each of them, so might as well be shared???

Another option is to repeat the IRQ handler in each of the files - it would be the same except for the IRQ.

What do you think?

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0