Update patch set 1
Patch Set 1: Do not submit

@Bill

Once you have built your EC with that CL, just type "ir357x" in the EC console and this should dump the full register space of the IR3571 (given it has been programmed with the same I2C as the IR3570)

I strongly advise not use the write functions (e.g "prog" or "write") since the IR chip has a very limited number of allowed flash write (ie 3) and is easy to mess up (e.g. locking it or changing its I2C address)

the "ir357x check" compared the current programming with the values provided by Srikanth to IR (they are pre-programming the parts before sending them to the factory).
By the way the current table is for the 3570, so it's probably not the right one. 

important note: AFAIK The IR357x can *only* be accessed when 3VS is on (ie in S0 or by manually setting ENABLE_VS to 1)
and it behaves differently when the CPU is ON and sending values on the SVID bus and when the CPU is OFF

Patch-set: 1
Label: Code-Review=-2
Label: Commit-Queue=0
Label: Verified=0