Update patch set 2
Patch Set 2: Abandoned

The component inputs that control the radios are WIFI_DISABLE and BT_DISABLE and are active-low (0v means disabled). The net names are WLAN_OFF# and BT_OFF#, so the EC's "ENABLE" names are correct (0v means disabled, +3v means enabled).

So we had it all right to begin with. Nevermind.

Patch-set: 2