Don't echo the NV recovery requests back to the BIOS. It knows.
When recovery is required, it will be because there's either a hardware pin
pulled somewhere, or because the recovery_reason is set in NVRAM. Coreboot
and U-Boot can see both of those, so the EC shouldn't make up a new reason.
If it does, it changes the original cause.

BUG=chrome-os-partner:9706
TEST=manual

Reset the EC using ESC+Power (Refresh+Power on EVT). At a root shell, run

  crossystem recovery_request=11
  reboot

When you see the Recovery screen, press TAB. It should say

  recovery_reason: 0x0b  We have no idea what this means

Prior to this fix, you'd see recovery_reason 2 instead, which is wrong.

Change-Id: Ie54185471927e7e829962d30bba9d142d593088f
Signed-off-by: Bill Richardson <wfrichar@chromium.org>