EC: snow: keyboard: fix incorrect index calculation with empty fifo
When kb_fifo_start is 0, the index calculation yields -1 because C got
the mod (%) operation arguably wrong.  By adding KB_FIFO_DEPTH to the
index before computing the mod with respect to KB_FIFO_DEPTH, we avoid
the negative case and (obviously) produce the same result in all other
cases.

BUG=chrome-os-partner:10247
TEST=saw incorrect state from kernel log before fix: could not repro after.

Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: I3a30c229dc9f762dd45203e842128811a24cf53f