Update patch set 1
Patch Set 1:

Going to let Katie make the call about whether we're going to include this.  

Luigi's WIP kernel patch ends up getting two interrupts for each event instead of 1, so he was seeing this a lot.  Should be much less often if that gets fixed, but this still seems like a reasonable fix to take if we're going to rev the EC.

Patch-set: 1