Modify Makefile to allow insertion of mock modules
When we need to mock modules in a unit test, we need a way to exclude
the actual modules. This CL enables exclusion of given source files.
For example, if we need to mock watchdog for test 'foo', we can add to
test/build.mk: "exclude-foo-core-y=watchdog.o", and then implement its own
watchdog and add "foo-y=foo_watchdog.o".

Signed-off-by: Vic Yang <victoryang@chromium.org>

BUG=chrome-os-partner:10356
TEST=Exclude watchdog for test hello and check compilation fails with
     "undefined reference to watchdog_task".

Change-Id: I4a0afb589a49dad7c4d6faf8926438085cdc46cf