Update patch set 4
Patch Set 4: Verified; No score

This is a trimmed down version of the VR driver to fix crosbug/p/10171 on EVT boards.
It is setting in-memory registers of the VR rather than re-flashing it because we are waiting for the updated flashing guide for the 3571 to do it correctly.

PTAL

Patch-set: 4
Reviewer: Gerrit User 1000016 <1000016@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Verified=+1