Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Rong this all looks great thanks you, but in terms of packaging I wonder if you could split the commits a little. I think it would be better for the i2c change to be a separate commit, and for the config file change to be a separate commit (and perhaps even just remove options rather than comment them out - add to the README if you want docs).

You have actually listed the three things in the commit message. If we split them then it is easier for people to see each logical change, and perhaps update or change it later.

Patch-set: 2
Reviewer: Gerrit User 1000294 <1000294@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=0