Update patch set 2
Patch Set 2: Verified; Ready

Correct -- host_command_reboot() is special because it must send a response to the host before rebooting. In the I2C case the host_buffer gets passed in as the "data" argument and is clobbered by host_send_response(). Then it is clobbered with EC_RES_SUCCESS (==0) when host_send_response() is called, so when the switch() statement is reached, p->cmd has been overwritten with 0 (EC_REBOOT_CANCEL).

Patch-set: 2
Label: Code-Review=0
Label: Verified=+1