Stop reporting keyboard recovery switch
This uses the new keyboard recovery EC event to determine when
keyboard recovery is requested.  It is paired with a change to the EC
to stop reporting keyboard recovery as a switch, since that didn't
provide a good mechanism for the AP to tell the EC to stop reporting
the switch value.

This REQUIRES an updated BIOS; without it you won't be able to get
into recovery mode manually.  You'll need to press space at the
developer screen, or use the debug console to request recovery.

BUG=chrome-os-partner:10034
TEST=manual

1. boot using power+refresh+esc.  See insert screen.
2. power button to power off
3. power button to power on.  Boots to Chrome OS (not insert screen).

Change-Id: I699ce004ed1190044170c4ea810b8969b40f523b
Signed-off-by: Randall Spangler <rspangler@chromium.org>