Print console messages for overtemp and shutdown
I have a system with bad temperature sensors and the EC is
constantly shutting the system down, but provides no
visible indication that it is doing so.

I have serious concerns about the EC behavior in this case
as it seems to be doing things it shouldn't.  However just
providing indication via the console about what it is doing
is at least useful for development and debug.

BUG=none
TEST=boot on system with bad temp sensors and see that
the EC indicates it is initiating a shutdown.

[1035.119825 x86 power state 3 = S0, in 0x5fff]
[critical threshold, shutting down]
[x86 power force shutdown]
[1035.198806 x86 power state 3 = S0, in 0x00ff]

Change-Id: I6beeb269a135bd8c245c0357670fe29648d48968
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>