Print console messages for overtemp and shutdown
I have a system with bad temperature sensors and the EC is
constantly shutting the system down, but provides no
visible indication that it is doing so.

I have serious concerns about the EC behavior in this case
as it seems to be doing things it shouldn't.  However just
providing indication via the console about what it is doing
is at least useful for development and debug.

BUG=none
TEST=boot on system with bad temp sensors and see that
the EC indicates it is initiating a shutdown.

[14.008340 critical temperature; shutting down]
[14.008660 x86 power force shutdown]
[14.009153 LPC RESET# asserted]

Change-Id: I6beeb269a135bd8c245c0357670fe29648d48968
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>