Update patch set 1
Patch Set 1:

Should the keyboard backlight come back automatically on resume, or should it wait for the kernel to turn it back on?

If it should come on automatically at the previous backlight value, use pwm_enable_keyboard_backlight(0) and also add a matching call to the resume handler, the way we do the fan.

If it should stay off until the kernel or BIOS explicitly turns it back on, LGTM as-is.

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=0