lpc: Use response size for response checksum
At present the protocol uses the parameter size for the response checksum.
This doesn't seem to make a lot of sense, since it is the response that
we want to check. Otherwise, the response could have an error and the
checksum will not tell us.

Change the response back to use the response size instead of param size.

BUG=chrome-os-partner:11317
TEST=manual:
'mkbp id' with new U-Boot supporting LPC
See that the command now succeeds since the checksums agree

Change-Id: I8409ca961ed53de73ad38eac53191c60457db4c6