Update patch set 2
Patch Set 2:

(In response to Simon's question: 

At the point in software sync where we're updating the EC, we only have the expected RW code buffer; we didn't think to provide its offset in the EC itself.  So it's easiest for the EC to tell us where it goes.

At OS auto-update time, this provides a quick check for whether the RO and RW sections have moved.  Won't be an issue for production, but may happen during development, and there's a factory test to verify the size of the WP region is correct.

Patch-set: 2