Stop keyboard scans from triggering ARM EC reset
The ARM EC was being rebooted when both the power and one of several other keys
were pressed. (LCtrl, Tab, Reload, t, [, ], y, Dim Screen and Mute) It should
only do this when the key combo PWR + Reload is pressed.

To fix it, keyboard scanning is disabled whenever the power button is
pressed.  It sets a flag indicating that scanning should be disabled,
and then any time the EC tries to select a scan column, it will just set
it to Hi-Z instead.

BUG=chrome-os-partner:10889
TEST=Pick one of the troublesome keys.  First press it, then quickly
press the power button.  Then press the power button followed
by the troublesome key.  Repeat this process several times for each key,
it should not reset the system.
Press power + reload, this should still reset the system.
Pressing and holding power should initiate a shutdown.

Change-Id: Ib60d2ebbb57eb8a3c135662514ec622c37a7eb07
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>