Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I think we were trying to get rid of the switches and move over to an host events.  I'm not 100% on that, but Randall would know.

Either way, I just tested this out on my dev. machine and it looks like it reintroduces the "sticky" recovery mode bug, though.

Also, it looks as if this is trying to fix the same problem as:
https://gerrit.chromium.org/gerrit/#/c/29016/

Patch-set: 1
Reviewer: Gerrit User 1001923 <1001923@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1
Label: Verified=0