Fixing lid power behavior -- shutdowns vs reboots
When the user is logged in a closes/opens the lid of the laptop the
lid_changed flag gets set, but never cleared.  Normally, it would get
cleared when it powered back on from being opened, but without powerm
running or is the machine is only suspended, then that never happens.
This fixes the problem by additionally clearing the flag whenever the
computer is powered down.  This forces the computer to not turn on
based on anything that happened to its lid before it was turned off,
which is the behavior you'd expect.

BUG=chrome-os-partner:12189
TEST=Log in to your Chomebook completely.  Close, then open the lid.
Shut the computer down.  It should stay off now.  Once it's back on
close the lid and confirm that it still suspends correctly.  Now
open the lid and make sure it turns back on.  Repeat these steps, but
kill powerm first.

Original-Change-Id: I2275b3125115b4eacc6a5d074978d7a1d51b0695
(cherry picked from commit 5e996478c9b41396771b97b0e2228d23f070ab45)

Change-Id: I0d425501be2446404ece6484c7a094705434e32d
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>