Fixing lid power behavior when powerm isnt working
If powerm isn't running, and the user closes then opens the lid, the
lid_changed flag gets set, but never cleared.  Normally, it would get
cleared when it turned back on from being opened, but without powerm
running then that never happens.  This fixes the problem by also
clearing the flag whenever the computer is powered down.  This forces
the computer to not turn on based on anything that happened to its lid
before it was turned off, which is the behavior you'd expect.

BUG=chrome-os-partner:12189
TEST=Stop the powerm task on your Chromebook.  Close, then open the lid.
Shut the computer down.  It should stay off now.  Once it's back on
(powerm should be running now) close the lid and confirm that it still
powers down.  Now open the lid and make sure it turns back on.

Change-Id: I2275b3125115b4eacc6a5d074978d7a1d51b0695
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>