Update patch set 2
Patch Set 2: (3 inline comments)

I'm not sure if this method of using some of the more significant bits of an error code is considered good practice here or not.  I haven't seen it done elsewhere, but it seemed like a nice simple solution to the problem of how to propagate the information back up with the error code itself.

Patch-set: 2