Lucas: Switching the i2c transmit code over to dma
There was an errata issues for the i2c on STMF100xx.  It specified that
not all guarantees apply to i2c on these chips if you are not using DMA
to load the data.  To prevent problems, I am converting the i2c code on
the EC for Lucas over to DMA.

Here the i2c's master functionality is retrofitted to use DMA
instead of polling to fill the i2c buffer.  The slave functionality
is still left in the old style for the time being, but will also be
converted soon.

BUG=chrome-os-partner:10901
TEST=From EC console, make sure that "battery" and "pmu" commands work.
They both use i2c, so if i2c had been broken they would fail.

Change-Id: I2fb80dcb68632938df1c9165ebd5a67cb5194451
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>