i2c: stm32: Increase tx timeout in slave mode from 10ms to 100ms
We have seen instances where the AP has interrupts disabled for a long
period of time (specifically when doing a lot of printk messages to
the console).  When this happens the host can't service i2c in < 10ms
(it needs an interrupt per byte) and we were getting a timeout.  We'll
increase the timeout to 100ms to avoid these problems.  Better to be
safe than sorry.

This timeout runs from the host command task so having the delay
shouldn't be a terrible thing (we're not running from an IRQ handler
or anything).

Only affected the timeout for slave mode specifically so as not to
affect any untested behavior.

BUG=chrome-os-partner:12123
TEST=With serial console enabled, run this in two different ssh
sessions:
a) while true; do flashrom -p internal:bus=i2c -r /tmp/ec.bin; done
b) while true; do /usr/local/lib/flimflam/test/connect-wifi GoogleGuest; done
...if flashrom reports success over and over again then this is good.

Change-Id: I7f32d5f1e4134896c857ee26f449a1fdd579d589
Signed-off-by: Doug Anderson <dianders@chromium.org>