Update patch set 1
Patch Set 1: Looks good to me, approved

I missed you only changed the timeout for the other case.

By the way, the 100ms *active* wait is very wasteful (CPU and power) we should sleep after some time (as we are doing for master) and maybe do exponential backoff for the sleep delay.
Please open a bug to fix that.

Patch-set: 1
Label: Code-Review=+2